requirements.txt
zeep
xmlsec
lxml
reportlab
qrcode
fastapi
uvicorn
psycopg2-binary>=2.9.9
